OPL-69: Move subscription to logic to separate subpackage
