+++ api_fhir_r4/models/__init__.py
@@ -1,4 +1,4 @@
 from api_fhir_r4.models.fhirModels import OperationOutcomeV2, UsageContextV2, CoverageV2, CoverageClassV2, \
     ContractSignerV2, ClaimV2, ClaimInsuranceV2, ClaimResponseV2, CoverageEligibilityRequestV2
 from api_fhir_r4.models.imisModelEnums import BundleType
-from api_fhir_r4.models.subscription import Subscription
+from api_fhir_r4.models.subscription import Subscription, SubscriptionNotificationResult
