+++ api_fhir_r4/serializers/__init__.py
@@ -90,5 +90,4 @@ from api_fhir_r4.serializers.communicationSerializer import CommunicationSeriali
 from api_fhir_r4.serializers.invoiceSerializer import InvoiceSerializer
 from api_fhir_r4.serializers.insuranceOrganisationSerializer import InsuranceOrganizationSerializer
 from api_fhir_r4.serializers.billSerializer import BillSerializer
-from api_fhir_r4.serializers.subscriptionSerializer import SubscriptionSerializer
 from api_fhir_r4.serializers.claimSerializer import ClaimSerializer
